requirements.txt
disnake>=2.8.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
requests>=2.31.0 